+++ src/openfermion/ops/_quadratic_hamiltonian_test.py
@@ -22,7 +22,7 @@ from openfermion.ops import (majorana_operator, normal_ordered,
 from openfermion.ops._quadratic_hamiltonian import antisymmetric_canonical_form
 from openfermion.transforms import get_fermion_operator, get_sparse_operator
 from openfermion.utils import get_ground_state
-from openfermion.tests._testing_utils import (random_antisymmetric_matrix,
+from openfermion.utils._testing_utils import (random_antisymmetric_matrix,
                                               random_hermitian_matrix)
